SJRK-239: overhauled dynamicValues passing for UI savingEnabled flag This commit updates the way dynamicValues are passed into templates in order to facilitate the injection of the UI-side savingEnabled flag from a single location. It also turns off editing, which should be noted.
To accomplish this, the renderTemplate invoker on all templateManager 
grades will now look for the dynamicValues collection in its own model 
rather than relying on it being passed in as an argument.

This comes with a number of advantages:
- tidier specification of new values; simply add a model value to a 
collection rather than overwriting an event listener in potentially 
several spots
- values can be merged across implementing grades
- less repetition (by virtue of eliminating the listener overwriting)
- the values can be operated upon or listened to by model relays or 
listeners

In addition to this work, a number of CSS and other changes were made 
which are covered by SJRK-237 but were necessary to make as part of 
adding a single flag to turn editing capabilities on or off.

Lastly, this commit removes some convenience features in a couple of 
templates which were used to make image path handling work even if the 
site were hosted in a subfolder. This will be dealt with in the future 
by SJRK-232.