SJRK-410: Prevent uploaded files from being removed from published stories during edit.
While editing a published story, if an uploaded file is modified, the original file would be deleted and the published story would look "broken" until the story edits were published. To prevent this, file deletion has been removed. We will need to investigate ways to re-enable file management for the users in future work.